+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pwr-rs"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 authors = ["Hamid R. K. Pishghadam <kaveh@riseup.net>"]
 license = "MIT OR Apache-2.0"
@@ -14,9 +14,9 @@ keywords = ["pwr", "sdk", "tool", "async"]
 rust-version = "1.63.0"
 
 [dependencies]
-secp256k1 = { version = "0.25.0", features = ["rand", "recovery"] }
+secp256k1 = { version = "0.29.0", features = ["rand", "recovery"] }
 rand = "0.8.5"
-reqwest = { version = "0.11.22", features = ["json"] }
+reqwest = { version = "0.12.5", features = ["json"] }
 serde = { version = "1.0.193", features = ["derive"] }
 serde_json = "1.0.108"
 sha3 = "0.10.8"
