+++ src/lib.rs
@@ -9,12 +9,15 @@
 //! # Create RPC
 //! ```ignore
 //! # use pwr_rs::wallet::Wallet;
-//! # use pwr_rs::rpc::RPC;
+//! # use pwr_rs::rpc::types::RPC;
 //! # let wallet = Wallet::random();
 //! let rpc = RPC::new("https://pwrrpc.pwrlabs.io/").unwrap();
 //! let balance = rpc.balance_of_address(&wallet.address()).await.unwrap();
 //! ```
 
 pub mod block;
+pub mod transaction;
+pub mod validator;
+pub mod delegator;
 pub mod rpc;
 pub mod wallet;
